Max b-tree mode support. Some tests don't work.

+++ include/memoria/containers/vector_tree/container/vtree_c_base.hpp
@@ -33,7 +33,7 @@ MEMORIA_BT_MODEL_BASE_CLASS_NO_CTOR_BEGIN(VTreeCtrBase)
     typedef typename Base::CtrShared                                            CtrShared;
 
     typedef LabeledTree<
-                FLabel<Short>,
+                FLabel<BigInt>,
                 VLabel<BigInt,
                     Granularity::Byte,
                     Indexed::Yes
